Added option to optimize for space over time

+++ transfer_images.sh
@@ -8,20 +8,22 @@ unset TAG
 unset UNBUNDLE
 unset IMAGELIST
 unset IMAGE
+unset SPACE
 USTMPDIR="/tmp"
 
-USAGE="Usage: $(basename $0) [-hr] [-l imagelist | -i image] [-d bundle.tar] [-t newtag] [-p tmpdir] [-u unbundle.tar]"
+USAGE="Usage: $(basename $0) [-hrs] [-l imagelist | -i image] [-d bundle.tar] [-t newtag] [-p tmpdir] [-u unbundle.tar]"
 HELP="
 -d bundle.tar - download container images and bundle them in a single tar file\n
 -h - help\n
 -i image - name of image to download\n
 -l imagelist - name of file that contains the list of images to download\n
 -p tmpdir\n
+-s - optimize for space over time
 -t newtag - retag images\n
 -u unbundle.tar - unbundle tar and load in container registry
 "
 
-while getopts d:hi:l:p:rt:u: c
+while getopts d:hi:l:p:srt:u: c
 do
 	case $c in
 	d) DOWNLOAD=${OPTARG};;
@@ -30,6 +32,7 @@ do
 	l) IMAGELIST=${OPTARG};;
 	r) RETAIN="true";;
 	p) USTMPDIR=${OPTARG};;
+	s) SPACE="true";;
 	t) TAG=${OPTARG};;
 	u) UNBUNDLE=${OPTARG};;
 	\?)	echo "${USAGE}"
@@ -68,6 +71,7 @@ fi
 if [ -n "${DOWNLOAD}" ];then
 	TMPDIR=$(mktemp -p ${USTMPDIR} -d --suffix=update_image)
 	for image in ${NEEDED_IMAGES};do
+		[[ ${image} =~ ^#.* ]] && continue
 		docker pull ${image}
 		if [ $? -ne 0 ];then
 			echo "Unable to download ${image}"
@@ -80,6 +84,9 @@ if [ -n "${DOWNLOAD}" ];then
 			rm -rf ${TMPDIR}
 			exit 1
 		fi
+		if [ -z "${RETAIN}" -a -n "${SPACE}" ];then
+			docker rmi ${image}
+		fi
 	done
 	tar -C ${TMPDIR} -cf ${DOWNLOAD} .
 	if [ $? -ne 0 ];then
@@ -87,7 +94,7 @@ if [ -n "${DOWNLOAD}" ];then
 		rm -rf ${TMPDIR}
 		exit 1
 	fi
-	if [ -z "${RETAIN}" ];then
+	if [ -z "${RETAIN}" -a -z "${SPACE}" ];then
 		for image in ${NEEDED_IMAGES};do
 			[[ ${image} =~ ^#.* ]] && continue
 			docker rmi ${image}
